LogFileLogger now only rotates logs above a certain size.

+++ src/main/java/com/controlj/logging/FileLogger.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019 Control-J Pty. Ltd. All rights reserved
+ * Copyright (c) 2019-2020 Control-J Pty. Ltd. All rights reserved
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,44 +18,57 @@
 package com.controlj.logging
 
 import com.controlj.logging.CJLog.logMsg
-import java.io.*
+import java.io.BufferedReader
+import java.io.File
+import java.io.FileReader
+import java.io.FileWriter
+import java.io.IOException
+import java.io.OutputStreamWriter
 import java.util.concurrent.BlockingQueue
 import java.util.concurrent.LinkedBlockingQueue
 
 /**
- * A logging [Destination] that saves to the supplied [logFile]. On each class creation any previous logfiles
- * will be rotated, up to a maximum of [MAX_FILES]. If the created file is "app.log" then older files will be
+ * A logging [Destination] that saves to the supplied [logFile]. On each class creation if the current logfile
+ * is greater in size than a limit, the previous logfiles
+ * will be rotated, up to a maximum of [maxFiles]. If the created file is "app.log" then older files will be
  * rotated to "app.log.1" etc.
  * Writing to files is carried out on a dedicated [Thread]
  *
  * @param logFile The file in which logs are to be saved
  */
 class FileLogger(val logFile: File) : Destination {
-    private val messageQueue: BlockingQueue<String> = LinkedBlockingQueue(MAX_QUEUE)
+    private val messageQueue: BlockingQueue<String> = LinkedBlockingQueue(100)
     private lateinit var logThread: Thread
+
+    /**
+     * The maximum number of rotated log files to keep
+     */
+    var maxFiles = 2
+
+    /**
+     * The maximum length of a log file before it is rotated.
+     */
+    var maxLength = 1_000_000
 
     init {
         logThread = Thread {
             try {
-                logMsg("Starting file logger, logfile is %s", logFile.name)
-                if(!logFile.parentFile.isDirectory)
+                logMsg("Starting file logger, logfile is ${logFile.name}")
+                if (!logFile.parentFile.isDirectory)
                     logFile.parentFile.mkdirs()
-                if (logFile.exists()) {
+                if (logFile.exists() && logFile.length() > maxLength) {
                     // rotate old log files so we can recover them
-                    val file = previous(MAX_FILES)
+                    val file = previous(maxFiles)
                     if (file.exists())
                         file.delete()
-                    for (i: Int in MAX_FILES - 1 downTo 1) {
+                    for (i: Int in maxFiles - 1 downTo 1) {
                         val f = previous(i)
-                        if (f.exists()) {
-                            val r = previous(i + 1)
-                            f.renameTo(r)
-                        }
+                        if (f.exists()) f.renameTo(previous(i + 1))
                     }
                     logFile.renameTo(previous(1))
                 }
                 val logWriter: OutputStreamWriter
-                logWriter = FileWriter(logFile, false)
+                logWriter = FileWriter(logFile, true)
                 while (!logThread.isInterrupted) {
                     val message = messageQueue.take()
                     logWriter.write(message)
@@ -99,7 +112,7 @@ class FileLogger(val logFile: File) : Destination {
     }
 
     override val previousLogfiles: List<File>
-            get() = (0 until MAX_FILES).map { previous(it) }.filter { it.exists() }.toList()
+        get() = (0 until maxFiles).map { previous(it) }.filter { it.exists() }.toList()
 
     override fun close() {
         logThread.interrupt()
@@ -108,13 +121,5 @@ class FileLogger(val logFile: File) : Destination {
     private fun previous(idx: Int): File {
         if (idx == 0) return logFile
         return File(logFile.parentFile, logFile.name + "." + idx)
-    }
-
-    companion object {
-        private const val MAX_QUEUE = 100
-        /**
-         * The maximum number of logfiles to be rotated and preserved.
-         */
-        var MAX_FILES = 4     // maximum number of files to keep
     }
 }
